# KAFKA-9039: Optimize ReplicaFetcher fetch path