@lbradstreet 2019-10-04
According to profiling, a majority of the rest of the time goes to looking up the logstartoffset for each partition on each pass (https://github.com/apache/kafka/pull/7443/files#diff-a8437f241a5ae585d5805ee769313080R252), and updating the incremental fetch session (). If we can drop those further, we may be able to save another 50-75%.