@guozhangwang 2019-10-11
4. Another wild thought: in `doWork` we can comment out `maybeTruncate` and see if that alone is taking up long time compared with the baseline, since in practice `maybeTruncate` should mostly be no-op, and if itself is taking lots of time we can also make another thorough pass on that function. I can do the benchmark taking from your branch if you think it's a good idea.