@lbradstreet 2019-10-11
> For 1) I've just updated the description to include the relative ratio of each optimization, it seems that [29fdd60](https://github.com/apache/kafka/commit/29fdd6094b828153c762f1d99b645481f7200cee) makes it less scalable with the num.partitions since it's only useful with small num.partitions.

Thanks! I'm assuming that you're OK with it still, because it still helps across the board? It just helps less for the high partition count case?