@lbradstreet 2019-10-11
> Is it always better to construct a new map `updated` than updating `next` directly?

It's a good question, and I'm not sure. The idea was that by making next smaller we could make the lookups progressively quicker as we process sessionsPartitions. It did seem to improve performance for our benchmark. I forgot to include the second optimization that it allows us to make https://github.com/apache/kafka/pull/7443/commits/f0bf23acd8198eae0faed67c2414301eccf04dc6. This helps reduce containsKey checks, and is worth keeping if we leave the change to keep the second map in. If not we will need to remove it again.