@lbradstreet 2019-10-15
> Is it always better to construct a new map `updated` than updating `next` directly?

Resolved by https://github.com/apache/kafka/pull/7443/commits/32e30ace5865e0a5b769cad626f64bcdf5091faf