@lbradstreet 2019-10-15
Build failures are due to broken trunk (https://github.com/apache/kafka/pull/7521)